bpf: Move go code in bpf/ into pkg/
