+++ cilium/cmd/bpf_policy_list.go
@@ -20,8 +20,8 @@ import (
 	"strconv"
 	"text/tabwriter"
 
-	"github.com/cilium/cilium/bpf/policymap"
 	"github.com/cilium/cilium/pkg/bpf"
+	"github.com/cilium/cilium/pkg/maps/policymap"
 	"github.com/cilium/cilium/pkg/policy"
 
 	"github.com/spf13/cobra"
