+++ daemon/config.go
@@ -19,10 +19,10 @@ import (
 	"net"
 	"sync"
 
-	"github.com/cilium/cilium/bpf/lxcmap"
 	"github.com/cilium/cilium/common/addressing"
 	"github.com/cilium/cilium/daemon/options"
 	"github.com/cilium/cilium/pkg/labels"
+	"github.com/cilium/cilium/pkg/maps/lxcmap"
 	"github.com/cilium/cilium/pkg/option"
 
 	etcdAPI "github.com/coreos/etcd/clientv3"
